feat(sks): abstract PT components
