+++ apps/protailwind/.eslintrc.js
@@ -1,5 +1,5 @@
 module.exports = {
   root: true,
   extends: ['custom'],
-  ignorePatterns: ['src/video/exercise/sandpack/parcel/*.js'],
+  ignorePatterns: ['src/exercise/sandpack/parcel/*.js'],
 }
