+++ apps/protailwind/package.json
@@ -120,9 +120,7 @@
     "react-rewards": "^2.0.3",
     "react-use": "^17.3.2",
     "react-wrap-balancer": "^0.2.4",
-    "refractor": "^4.5.0",
     "shortid": "^2.2.16",
-    "speakingurl": "^14.0.1",
     "superjson": "=1.11.0",
     "tailwindcss-radix": "^2.8.0",
     "uuid": "^8.3.2",
@@ -150,7 +148,6 @@
     "@types/react": "^18.0.15",
     "@types/react-dom": "^18.0.6",
     "@types/shortid": "^0.0.29",
-    "@types/speakingurl": "^13.0.3",
     "@types/uuid": "^8.3.4",
     "autoprefixer": "^10.4.2",
     "eslint": "^8.19.0",
@@ -158,7 +155,6 @@
     "msw": "^0.35.0",
     "postcss": "^8.4.14",
     "prettier-plugin-tailwindcss": "^0.2.5",
-    "react-refractor": "^2.1.7",
     "tailwindcss": "^3.3.0",
     "@tailwindcss/typography": "^0.5.9",
     "tailwind-scrollbar": "^3.0.0",
