+++ apps/protailwind/src/components/gitpod-link.tsx
@@ -1,19 +1,19 @@
 import {useLesson} from '@skillrecordings/skill-lesson/hooks/use-lesson'
-import Icon from 'components/icons'
+import {Icon} from '@skillrecordings/react'
 import {useRouter} from 'next/router'
 import {track} from 'utils/analytics'
 import {trpc} from 'trpc/trpc.client'
 
-const GitHubLink = () => {
+const GitpodLink = () => {
   const router = useRouter()
   const {lesson, module} = useLesson()
   const {data: resources} = trpc.resources.byExerciseSlug.useQuery({
     slug: router.query.lesson as string,
     type: lesson._type,
   })
-  const github = resources?.github
+  const gitpod = resources?.gitpod
 
-  if (!github) {
+  if (!gitpod) {
     return null
   }
 
@@ -21,21 +21,21 @@ const GitHubLink = () => {
     <div className="flex items-center gap-2">
       <a
         onClick={() => {
-          track('clicked github code link', {
+          track('clicked gitpod code link', {
             lesson: lesson.slug,
             module: module.slug.current,
             moduleType: module.moduleType,
             lessonType: lesson._type,
           })
         }}
-        href={github.url}
+        href={gitpod.url}
         target="_blank"
         rel="noopener noreferrer"
-        className="inline-flex items-center gap-2 rounded-lg bg-gray-800 py-2 px-4 text-lg font-medium text-white transition hover:bg-gray-900"
+        className="inline-flex items-center gap-2 rounded-lg bg-gray-800 px-4 py-2 text-lg font-medium text-white transition hover:bg-gray-900"
       >
-        <Icon name="Github" size="24" />
+        <Icon name="Gitpod" size="24" />
         <div>
-          <p className="font-semibold">Code</p>
+          <p className="font-semibold">Gitpod</p>
           {/* <p className="font-mono text-sm text-gray-400">/{openFile}</p> */}
         </div>
       </a>
@@ -43,4 +43,4 @@ const GitHubLink = () => {
   )
 }
 
-export default GitHubLink
+export default GitpodLink
