+++ apps/total-typescript/package.json
@@ -104,15 +104,12 @@
     "react-hook-form": "^7.29.0",
     "react-hot-toast": "^2.2.0",
     "react-markdown": "^8.0.4",
-    "react-refractor": "^2.1.7",
     "react-rewards": "^2.0.3",
     "react-scroll": "^1.8.6",
     "react-use": "^17.3.2",
     "react-use-clipboard": "^1.0.7",
     "react-wrap-balancer": "^0.2.4",
-    "refractor": "^4.5.0",
     "shortid": "^2.2.16",
-    "speakingurl": "^14.0.1",
     "superjson": "=1.11.0",
     "tailwindcss-radix": "^2.8.0",
     "uuid": "^8.3.2",
@@ -143,7 +140,6 @@
     "@types/react-dom": "^18.0.6",
     "@types/react-scroll": "^1.8.3",
     "@types/shortid": "^0.0.29",
-    "@types/speakingurl": "^13.0.3",
     "@types/uuid": "^8.3.4",
     "@types/ws": "^8.5.3",
     "autoprefixer": "^10.4.2",
