+++ apps/total-typescript/src/styles/globals.css
@@ -8,12 +8,15 @@
 @import './prism-theme.css';
 @import './pricing.css';
 @import './purchase-transfer.css';
+@import './portable-text.css';
+
 @import '../video/module-lesson-list/styles.css';
 @import '../video/video-overlays/styles.css';
 @import '../video/lesson-completion-toggle/styles.css';
 @import '../video/lesson-description/styles.css';
 @import '../video/lesson-title/styles.css';
 @import '../video/video-transcript/styles.css';
+@import '../video/github-link/styles.css';
 
 @tailwind utilities;
 @import './focus-visible.css';
