+++ apps/total-typescript/src/templates/article-template.tsx
@@ -11,7 +11,8 @@ import {
 } from 'components/primary-newsletter-cta'
 import {format} from 'date-fns'
 import {ArticleTeaser} from 'pages/articles'
-import PortableTextComponents from 'video/portable-text'
+import {portableTextComponents} from '@skillrecordings/react'
+import Spinner from 'components/spinner'
 
 type ArticleTemplateProps = {
   article: Article
@@ -55,7 +56,7 @@ const ArticleTemplate: React.FC<ArticleTemplateProps> = ({
         images={[image || '']}
         description={articleDescription}
       />
-      <header className="relative z-10 flex w-full flex-col items-center justify-center  px-5 pt-24 pb-8 sm:pt-36 sm:pb-10">
+      <header className="relative z-10 flex w-full flex-col items-center justify-center  px-5 pb-8 pt-24 sm:pb-10 sm:pt-36">
         <div className="relative z-10 mx-auto flex w-full max-w-3xl flex-col">
           <time dateTime={_createdAt} className="pb-4 text-gray-300">
             {format(new Date(_createdAt), 'MMM dd, y')}
@@ -89,8 +90,11 @@ const ArticleTemplate: React.FC<ArticleTemplateProps> = ({
         )}
       </header>
       <main className="relative z-10 pt-5">
-        <div className="prose relative z-10 mx-auto w-full max-w-3xl px-5 prose-p:text-gray-300 prose-a:text-cyan-300 prose-a:transition hover:prose-a:text-cyan-200 sm:prose-lg sm:prose-pre:-mx-5 md:prose-xl">
-          <PortableText value={body} components={PortableTextComponents} />
+        <div className="prose relative z-10 mx-auto w-full max-w-3xl px-5 sm:prose-lg md:prose-xl prose-p:text-gray-300 prose-a:text-cyan-300 prose-a:transition hover:prose-a:text-cyan-200 sm:prose-pre:-mx-5">
+          <PortableText
+            value={body}
+            components={portableTextComponents({loadingIndicator: <Spinner />})}
+          />
           <div className="flex w-36 -rotate-6 gap-2 pt-10 text-gray-400">
             —
             <Image
