+++ packages/react/package.json
@@ -41,7 +41,11 @@
     "react-use-clipboard": "^1.0.7",
     "react-markdown": "^6.0.3",
     "react-syntax-highlighter": "^15.4.4",
-    "yup": "^0.32.11"
+    "yup": "^0.32.11",
+    "react-refractor": "^2.1.7",
+    "refractor": "^4.5.0",
+    "speakingurl": "^14.0.1",
+    "@portabletext/react": "^1.0.6"
   },
   "devDependencies": {
     "@skillrecordings/scripts": "workspace:*",
@@ -67,7 +71,9 @@
     "react-dom": "18.2.0",
     "sass": "^1.43.4",
     "typescript": "^4.9.4",
-    "xstate": "=4.23.1"
+    "xstate": "=4.23.1",
+    "@portabletext/types": "^1.0.3",
+    "@types/speakingurl": "^13.0.3"
   },
   "peerDependencies": {
     "@xstate/react": "=1.5.1",
