+++ packages/react/src/index.ts
@@ -21,3 +21,6 @@ export * from './hooks/use-debounce'
 export * from './components'
 export * from './utils'
 export * from './hooks'
+export {Icon, type IconNames} from './icons'
+export {portableTextComponents} from './portable-text'
+export {TableOfContents} from './portable-text/table-of-contents'
